feat: optimized layer caching and image size
Operations inside the Dockerfiles have been reordered to ensure
dependencies are installed first in both images and all redundant files
have been cleaned up after the operation. The Python image now builds
from Alpine, testing is required to ensure compatilibity.

This also moves an environment variable from the general one to the
Dockerfile itself to prevent polluting the production environment as it
is not needed there.