feat: add compatibility with versioned images and local models
This commit changes the default tag pulled by the images from 'main' to
'latest', which will be automatically given to the most recent image
built from the latest release, while also enabling it to be changed to a
specific version by specifying it in a .env file.

It also creates the new models directory to store the Spacy files for
the extra languages, and modifies the PYTHONPATH environment variable to
append this folder to it.