fix: commenting in and out is no longer needed and silenced warning
This commit makes it so that users of Apple Silicon no longer need to
comment out a line to run the python container nor remember to comment
it back in when updating; the process is now handled by an enviroment
variable like all the other deployment settings.

It also silences a warning when trying to append a directory to the
PYTHONPATH environment variable inside the container because the
variable is not set up by default. Now it is simply overwritten.