fix: trying to remove a missing folder is not an error anymore
This commit makes it so that trying to remove the directory with the
local python packages for the language models no longer returns an error
status but a successful one since operation was performed as intended.

This commit also ensures that removal is retried up to four additional
times to circumvent an error that would occasionally impede removel on
the first attempt with no noticeable effects on subsequent ones.