feat: remove version and migrate to named volume for shared folder
This commit moves from using simple mounts to using a named volume,
which copies the contents present at the location in the container into
the bound location so that now an empty folder is all that is needed for
the deployment. No subfolders needed so that cloning a branch of the
repo is optional.

An option to bind the volume to a different location was also added for
those users wanting the extra flexibility.

This commit also removes the 'version' section from the compose file
since it exists only for backwards compatibility with the V1 of the
compose specification which was superseeded 6 years ago.