feat: streamlined python container build and config
This commit switches the base image from ubuntu to python:slim as it
makes for a more sensible default that simplifies the build process
without compromising the final size.

It also moves the setup of the environment variable for the PYTHONPATH
and the command to run the main process inside the container to the
image itself, where they belong as they are not configurable and any
change would break the service.